Operator-Script/requirements.txt
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
exceptiongroup==1.2.2
firebase-admin==6.6.0
firestore==0.0.8
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
iniconfig==2.0.0
iso8601==2.1.0
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
pytest==8.3.3
pytz==2024.2
requests==2.32.3
rsa==4.9
tomli==2.2.1
uritemplate==4.1.1
urllib3==2.2.3
